+++ Indicators/OptionGreekIndicatorBase.cs
@@ -23,57 +23,12 @@ namespace QuantConnect.Indicators
     /// <summary>
     /// To provide a base class for option greeks indicator
     /// </summary>
-    public abstract class OptionGreeksIndicatorBase : IndicatorBase<IndicatorDataPoint>, IIndicatorWarmUpPeriodProvider
+    public abstract class OptionGreeksIndicatorBase : OptionIndicatorBase
     {
-        /// <summary>
-        /// Option's symbol object
-        /// </summary>
-        protected readonly Symbol _optionSymbol;
-
-        /// <summary>
-        /// Underlying security's symbol object
-        /// </summary>
-        protected Symbol _underlyingSymbol => _optionSymbol.Underlying;
-
         /// <summary>
         /// Cache of the current value of the greek
         /// </summary>
         protected decimal _greekValue;
-
-        /// <summary>
-        /// Option pricing model used to calculate greeks
-        /// </summary>
-        protected OptionPricingModelType _optionModel;
-
-        /// <summary>
-        /// Risk-free rate model
-        /// </summary>
-        protected readonly IRiskFreeInterestRateModel _riskFreeInterestRateModel;
-
-        /// <summary>
-        /// Gets the expiration time of the option
-        /// </summary>
-        public DateTime Expiry => _optionSymbol.ID.Date;
-
-        /// <summary>
-        /// Gets the option right (call/put) of the option
-        /// </summary>
-        public OptionRight Right => _optionSymbol.ID.OptionRight;
-
-        /// <summary>
-        /// Gets the strike price of the option
-        /// </summary>
-        public decimal Strike => _optionSymbol.ID.StrikePrice;
-
-        /// <summary>
-        /// Gets the option style (European/American) of the option
-        /// </summary>
-        public OptionStyle Style => _optionSymbol.ID.OptionStyle;
-
-        /// <summary>
-        /// Risk Free Rate
-        /// </summary>
-        public Identity RiskFreeRate { get; set; }
 
         /// <summary>
         /// Gets the implied volatility of the option
@@ -81,54 +36,33 @@ namespace QuantConnect.Indicators
         public IndicatorBase<IndicatorDataPoint> ImpliedVolatility { get; }
 
         /// <summary>
-        /// Gets the option price level
-        /// </summary>
-        public IndicatorBase<IndicatorDataPoint> Price { get; }
-
-        /// <summary>
-        /// Gets the underlying's price level
-        /// </summary>
-        public IndicatorBase<IndicatorDataPoint> UnderlyingPrice { get; }
-
-        /// <summary>
-        /// Initializes a new instance of the ImpliedVolatility class
+        /// Initializes a new instance of the OptionGreeksIndicatorBase class
         /// </summary>
         /// <param name="name">The name of this indicator</param>
         /// <param name="option">The option to be tracked</param>
         /// <param name="riskFreeRateModel">Risk-free rate model</param>
         /// <param name="period">The lookback period of historical volatility</param>
-        /// <param name="optionModel">The option pricing model used to estimate IV</param>
+        /// <param name="optionModel">The option pricing model used to estimate the Greek</param>
+        /// <param name="ivModel">The option pricing model used to estimate IV</param>
         protected OptionGreeksIndicatorBase(string name, Symbol option, IRiskFreeInterestRateModel riskFreeRateModel, int period = 2,
             OptionPricingModelType optionModel = OptionPricingModelType.BlackScholes, OptionPricingModelType? ivModel = null)
-            : base(name)
+            : base(name, option, riskFreeRateModel, period, optionModel)
         {
-            var sid = option.ID;
-            if (!sid.SecurityType.IsOption())
-            {
-                throw new ArgumentException("ImpliedVolatility only support SecurityType.Option.");
-            }
-
-            _optionSymbol = option;
-            _riskFreeInterestRateModel = riskFreeRateModel;
-            _optionModel = optionModel;
             ivModel = ivModel ?? optionModel;
-
-            RiskFreeRate = new Identity(name + "_RiskFreeRate");
             ImpliedVolatility = new ImpliedVolatility(name + "_IV", option, riskFreeRateModel, period, (OptionPricingModelType)ivModel);
-            Price = new Identity(name + "_Close");
-            UnderlyingPrice = new Identity(name + "_UnderlyingClose");
-
+            
             WarmUpPeriod = period;
         }
 
         /// <summary>
-        /// Initializes a new instance of the ImpliedVolatility class
+        /// Initializes a new instance of the OptionGreeksIndicatorBase class
         /// </summary>
         /// <param name="name">The name of this indicator</param>
         /// <param name="option">The option to be tracked</param>
         /// <param name="riskFreeRateModel">Risk-free rate model</param>
         /// <param name="period">The lookback period of historical volatility</param>
-        /// <param name="optionModel">The option pricing model used to estimate IV</param>
+        /// <param name="optionModel">The option pricing model used to estimate the Greek</param>
+        /// <param name="ivModel">The option pricing model used to estimate IV</param>
         protected OptionGreeksIndicatorBase(string name, Symbol option, PyObject riskFreeRateModel, int period = 2,
             OptionPricingModelType optionModel = OptionPricingModelType.BlackScholes, OptionPricingModelType? ivModel = null)
             : this(name, option, RiskFreeInterestRateModelPythonWrapper.FromPyObject(riskFreeRateModel), period, optionModel, ivModel)
@@ -136,13 +70,14 @@ namespace QuantConnect.Indicators
         }
 
         /// <summary>
-        /// Initializes a new instance of the ImpliedVolatility class
+        /// Initializes a new instance of the OptionGreeksIndicatorBase class
         /// </summary>
         /// <param name="name">The name of this indicator</param>
         /// <param name="option">The option to be tracked</param>
         /// <param name="riskFreeRate">Risk-free rate, as a constant</param>
         /// <param name="period">The lookback period of historical volatility</param>
-        /// <param name="optionModel">The option pricing model used to estimate IV</param>
+        /// <param name="optionModel">The option pricing model used to estimate the Greek</param>
+        /// <param name="ivModel">The option pricing model used to estimate IV</param>
         protected OptionGreeksIndicatorBase(string name, Symbol option, decimal riskFreeRate = 0.05m, int period = 2,
             OptionPricingModelType optionModel = OptionPricingModelType.BlackScholes, OptionPricingModelType? ivModel = null)
             : this(name, option, new ConstantRiskFreeRateInterestRateModel(riskFreeRate), period, optionModel, ivModel)
@@ -155,13 +90,7 @@ namespace QuantConnect.Indicators
         public override bool IsReady => Samples >= 2 && Price.Current.Time == UnderlyingPrice.Current.Time;
 
         /// <summary>
-        /// Required period, in data points, for the indicator to be ready and fully initialized.
-        /// </summary>
-        public int WarmUpPeriod { get; }
-
-        /// <summary>
-        /// Computes the next value of the following sub-indicators from the given state:
-        /// StandardDeviation, MiddleBand, UpperBand, LowerBand, BandWidth, %B
+        /// Computes the next value of the option greek indicator
         /// </summary>
         /// <param name="input">The input given to the indicator</param>
         /// <returns>The input is returned unmodified.</returns>
@@ -193,21 +122,18 @@ namespace QuantConnect.Indicators
             return _greekValue;
         }
 
-        // Calculate the Delta of the option
+        // Calculate the greek of the option
         protected virtual decimal CalculateGreek(DateTime time)
         {
             throw new NotImplementedException("'CalculateGreek' method must be implemented");
         }
 
         /// <summary>
-        /// Resets this indicator and all sub-indicators (StandardDeviation, LowerBand, MiddleBand, UpperBand, BandWidth, %B)
+        /// Resets this indicator and all sub-indicators
         /// </summary>
         public override void Reset()
         {
-            RiskFreeRate.Reset();
             ImpliedVolatility.Reset();
-            Price.Reset();
-            UnderlyingPrice.Reset();
             base.Reset();
         }
     }
