+++ Tests/Algorithm/AlgorithmIndicatorsTests.cs
@@ -91,64 +91,5 @@ namespace QuantConnect.Tests.Algorithm
             // Our interest rate provider should have been called once
             interestRateProviderMock.Verify(x => x.GetInterestRate(reference), Times.Once);
         }
-
-        [Test]
-        public void IVIndicatorUsesAlgorithmsRiskFreeRateModelSetAfterIndicatorRegistration()
-        {
-            // Register indicator
-            var iv = _algorithm.IV(_option);
-
-            // Setup risk free rate model
-            var interestRateProviderMock = new Mock<IRiskFreeInterestRateModel>();
-            var reference = new DateTime(2023, 11, 21, 10, 0, 0);
-            interestRateProviderMock.Setup(x => x.GetInterestRate(reference)).Verifiable();
-
-            // Update indicator
-            iv.Update(new IndicatorDataPoint(_option, reference, 30m));
-            iv.Update(new IndicatorDataPoint(Symbols.SPY, reference, 300m));
-
-            // Our interest rate provider shouldn't have been called yet since it's hasn't been set to the algorithm
-            interestRateProviderMock.Verify(x => x.GetInterestRate(reference), Times.Never);
-
-            // Set the interest rate provider to the algorithm
-            _algorithm.SetRiskFreeInterestRateModel(interestRateProviderMock.Object);
-
-            // Update indicator
-            iv.Update(new IndicatorDataPoint(_option, reference, 30m));
-            iv.Update(new IndicatorDataPoint(Symbols.SPY, reference, 300m));
-
-            // Our interest rate provider should have been called once by each update
-            interestRateProviderMock.Verify(x => x.GetInterestRate(reference), Times.Exactly(2));
-        }
-
-        [Test]
-        public void DeltaIndicatorUsesAlgorithmsRiskFreeRateModelSetAfterIndicatorRegistration()
-        {
-            // Register indicator
-            var delta = _algorithm.Delta(_option);
-
-            // Setup risk free rate model
-            var interestRateProviderMock = new Mock<IRiskFreeInterestRateModel>();
-            var reference = new DateTime(2023, 8, 1, 10, 0, 0);
-            interestRateProviderMock.Setup(x => x.GetInterestRate(reference)).Verifiable();
-
-            // Update indicator
-            delta.Update(new IndicatorDataPoint(_option, reference, 30m));
-            delta.Update(new IndicatorDataPoint(Symbols.SPY, reference, 300m));
-
-            // Our interest rate provider shouldn't have been called yet since it's hasn't been set to the algorithm
-            interestRateProviderMock.Verify(x => x.GetInterestRate(reference), Times.Never);
-
-            // Set the interest rate provider to the algorithm
-            _algorithm.SetRiskFreeInterestRateModel(interestRateProviderMock.Object);
-
-            // Update indicator
-            delta.Update(new IndicatorDataPoint(_option, reference.AddDays(1), 30m));
-            delta.Update(new IndicatorDataPoint(Symbols.SPY, reference.AddDays(1), 300m));
-
-            // Our interest rate provider should have been called once by each update
-            // 1 for delta, 2 for IV
-            interestRateProviderMock.Verify(x => x.GetInterestRate(reference.AddDays(1)), Times.Exactly(3));
-        }
     }
 }
